Updated README with formatting and Unity Peoples
I should've put this into one of the last 3 commits I've made on this Readme.
The "Unity Peoples" above is referencing the new table I've made of job positions, full names, and GitHub usernames.  Those together took the better part of a half hour to collect from a bunch of different sources.  Use that table wisely.