A Simple (Minecrafty) Sword Script via Sine Wave
I employed a Sine Wave to get the sword to move, and, well, it looks like Minecraft.   A LOT like Minecraft.
If I had to name one concern, it would be that the attack times (when trying to attack as fast as possible) seem a little inconsistent.  I've no idea why.