Dialogue now reads from file (yay)
The new system allows for scripts to be written that specify which key from each object's dialogue dictionary should be chosen based on the current state.  For example, an object could choose to use the "give_quest" entry of the database before giving a quest, then using the "continue_quest" when that player returns without completing his duties.
I also changed the "Talk" button from whatever it was (which wasn't working on my computer) to click or "Q".