The Dialogue UI is now fully functional, if terrifying.
It works.  It's an eyesore, but it works from JSON like a dream.  Thing is, you can't actually... do things with these conversations.  Until we add some sort of "action" functionality to our JSON definitions, all the Dialogue system will be good for is Small talk.  Small talk is boring.