Un-wrong-unity-version-afied the project, and fixed cryptic Dialogue system bug.
After a fair bit of debugging, I discovered that the error (a cast exception) was being thrown inside the Messenger system, since it was trying to cast a method with one argument (Dialogue) to a listener of two (Dialogue and Speaker), as per the changes we made at the last Unity class.  Just switch the Event Producer to give both necessary arguments, and bam, all is well.