Finished off Quest System with JSON.  It is missing some goals and rewards, and I have yet to test each of those two that do exist.
Still, a quite a bit of work has been done.  I've also fixed some little issues, such as the Converer part of the Dialogue system (it wasn't using the WorldInteraction system).  I also merged the JSON file reading functionalities of QuestManager and DialogueManager into a single JsonFileReader base class.
I also changed the Unity Scene file in a couple of ways, such as removing a Sprite Check thing that we don't use and which slows down the play button for 6.7 seonds (on my computer)